fastapi_sse/requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
gunicorn>=23.0.0
redis>=5.0.0
celery[redis]>=5.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.22
httpx>=0.28.0
prometheus-client>=0.19.0
structlog>=24.1.0
python-json-logger>=2.0.7
orjson>=3.11.6
tenacity>=8.2.0
pyjwt>=2.12.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-redis>=0.43b0
opentelemetry-propagator-b3>=1.22.0
